functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.0.0
ai-edge-litert>=1.0.1
numpy>=1.24.0
pillow>=10.0.0
